error override - match against spectest expected error message exactly
